# api/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_Gtport_DPmodule_internal_service.MapDataDTO:
    properties:
      groups:
        description: с координатами — на карту
        items:
          $ref: '#/definitions/github_com_Gtport_DPmodule_internal_service.MapGroupDTO'
        type: array
      no_coords:
        description: станция без координат в справочнике — боковой список
        items:
          $ref: '#/definitions/github_com_Gtport_DPmodule_internal_service.MapGroupDTO'
        type: array
      targets:
        description: терминалы для чипов (правило унификации)
        items:
          $ref: '#/definitions/github_com_Gtport_DPmodule_internal_service.TargetDTO'
        type: array
      total:
        description: групп всего
        type: integer
      vagons:
        description: вагонов всего в снимке
        type: integer
    type: object
  github_com_Gtport_DPmodule_internal_service.MapGroupDTO:
    properties:
      arrived:
        description: все вагоны на станции назначения (9/10/12)
        type: boolean
      broshen:
        description: есть вагон-5
        type: boolean
      doroga_oper:
        type: string
      index:
        type: string
      key:
        description: trainKey — им же ходят ручки вагонов и пометки
        type: string
      lat:
        description: nil → группа уходит в no_coords
        type: number
      lon:
        type: number
      mark_color:
        description: 'пометка диспетчера: info_2'
        type: string
      mark_text:
        description: 'пометка диспетчера: info_1'
        type: string
      naznach:
        description: терминал-большинство → цвет маркера
        type: string
      naznach_list:
        description: все терминалы группы → фильтр чипами
        items:
          type: string
        type: array
      oper_s:
        type: string
      plan_jd:
        type: string
      prog_jd:
        description: непустой = «ходовой», пустой = «отцепка»
        type: string
      rasst:
        description: остаток до станции назначения, км
        type: integer
      stan_nazn:
        type: string
      station_oper:
        type: string
      status:
        description: большинство по вагонам (как «Поезда»)
        type: integer
      time_op:
        type: string
      vagon_count:
        type: integer
    type: object
  github_com_Gtport_DPmodule_internal_service.MapWagonDTO:
    properties:
      freight:
        description: freight_exact_name, пусто — cargo_s
        type: string
      gruzpol_s:
        type: string
      id:
        description: id рейса — для «Истории движения вагона» по ПКМ
        type: string
      invoice:
        type: string
      naznach:
        type: string
      npp_vag:
        type: integer
      owner:
        type: string
      status:
        type: integer
      vagon:
        type: string
      ves:
        type: number
    type: object
  github_com_Gtport_DPmodule_internal_service.MapWagonsDTO:
    properties:
      key:
        type: string
      wagons:
        items:
          $ref: '#/definitions/github_com_Gtport_DPmodule_internal_service.MapWagonDTO'
        type: array
    type: object
...
  github_com_Gtport_DPmodule_internal_service.RearrApplyResult:
    properties:
      forecast_computed:
        description: пересчитан ход (Stage 3)
        type: integer
      prog_computed:
        description: пересчитан прогноз порта (Stage 4)
        type: integer
      selected:
        description: вагонов было выбрано
        type: integer
      updated:
        description: вагонов изменено
        type: integer
    type: object
...
  github_com_Gtport_DPmodule_internal_service.TargetDTO:
    properties:
      color:
        description: ports.color — цвет терминала в шапках таблиц
        type: string
      name:
        description: NameS терминала (значение naznach)
        type: string
      station:
        description: имя причальной станции терминала
        type: string
      station_code:
        description: 4-значный код станции (= code4_stan_nazn вагона)
        type: string
    type: object
...
  internal_handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: IQPort microservice template.
  title: IQPort Service API
  version: "1.0"
paths:
...
  /api/v1/dislocation/map:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Gtport_DPmodule_internal_service.MapDataDTO'
      security:
      - BearerAuth: []
      summary: 'Карта: все группы «поезд на станции» с координатами + терминалы для
        чипов'
      tags:
      - dislocation
  /api/v1/dislocation/map/mark:
    post:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Gtport_DPmodule_internal_service.RearrApplyResult'
      security:
      - BearerAuth: []
      summary: 'Карта: пометка выбранных групп (текст+цвет в info_1/info_2; оба пустых
        = снять)'
      tags:
      - dislocation
  /api/v1/dislocation/map/wagons:
    get:
      parameters:
      - description: ключ группы (key из /dislocation/map)
        in: query
        name: key
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Gtport_DPmodule_internal_service.MapWagonsDTO'
      security:
      - BearerAuth: []
      summary: 'Карта: вагоны одной группы (drill-down из попапа маркера)'
      tags:
      - dislocation
...
  /api/v1/map/tiles/{z}/{x}/{y}:
    get:
      parameters:
      - description: зум (0–11; сплошное покрытие кэша — 4–8)
        in: path
        name: z
        required: true
        type: integer
      - description: колонка тайла
        in: path
        name: x
        required: true
        type: integer
      - description: строка тайла
        in: path
        name: "y"
        required: true
        type: integer
      produces:
      - image/png
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: тайла нет в кэше
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Тайл подложки карты из кэша OSM (публичная, без JWT)
      tags:
      - map
...
securityDefinitions:
  BearerAuth:
    description: 'Keycloak JWT. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"